% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ikfdnm_mcmc}
\alias{ikfdnm_mcmc}
\title{Perform MCMC sampling for integrated known-fate, dynamic N-mixture model}
\usage{
ikfdnm_mcmc(kfdnm_lik, prior_dist, sample_N = FALSE, sample_SG = FALSE,
  burn, iter, block, par)
}
\arguments{
\item{kfdnm_lik}{A likelihood function created with the \code{\link{make_ikfdnm_likelihood}} function}

\item{prior_dist}{A user defined function that evaluates to the log-prior distribution of the parameter vector}

\item{sample_N}{Logical indicating that the abundance process should be sampled.}

\item{sample_SG}{Logical indicating wheather the survival (S) and recruitement process should be sampled.}

\item{burn}{The number of iterations for burnin of the MCMC algorithm}

\item{iter}{The number of MCMC iterations}

\item{block}{Block size for adaptation of proposal variance.}

\item{par}{Initial parameter value. If not provided a optimization is executed first to obtain
realistic initial values.}
}
\value{
A list containg the MCMC sample of parameters, abundance, and dynamic abundance components.
}
\description{
This function used a KFDNM likelihood function, as well as, a user defined prior distribution
for the parameters and uses it to perform MCMC sampling of the parameter vector.
}

