\name{kfdnm-package}
\alias{kfdnm-package}
\alias{kfdnm}
\docType{package}
\title{
Integrated known-fate, dynamic N-mixture model fitting
}
\description{
This package provides functions for fitting the integrated known-fate dynamic N-mixture model using Markov Chain Monte Carlo and maximum likelihood.
}
\details{
\tabular{ll}{
Package: \tab kfdnm\cr
Type: \tab Package\cr
Version: \tab 0.001\cr
Date: \tab 2014-10-24\cr
License: \tab Unlimited\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Devin S. Johnson

Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{known-fate, N-mixture}
