% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_ikfdnm_likelihood}
\alias{make_ikfdnm_likelihood}
\title{Create IKFDNM model likehood function}
\usage{
make_ikfdnm_likelihood(dnm_survival = ~1, dnm_recruit = ~1, dnm_det = ~1,
  kf_survival_effects = NULL, fixed_list = NULL, data, N_max)
}
\description{
Create IKFDNM model likehood function
}

