% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_group}
\alias{sim_group}
\title{Simulate group data for analysis with known-fate dynamic N-mixture models}
\usage{
sim_group(num_KF, num_years, num_surveys, recruit_rate, init_rate,
  annual_survival, annual_survival_KF, perfect_survey_rate, detection)
}
\description{
This function generates simulated data in the correct form for use in the kfdnm package
}

