% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{kfdnm-package}
\alias{kfdnm}
\alias{kfdnm-package}
\title{Fit integrated known-fate, dynamic N-mixture models}
\description{
The \code{kfdnm} package provide some basic functions for fitting itegrated known-fate, dynamic N-mixture models.
The functions can be used for both maximum-likelihood and Bayesian (via MCMC) inference. A Hidden Markov Model (HMM) formulation allows efficient sampling
of the abundance process in an MCMC or calculation of the likelihood for optimization.


\tabular{ll}{
Package: \tab crawl\cr
Type: \tab Package\cr
Version: \tab 1.901\cr
Date: \tab October 23, 2014\cr
License: \tab Unlimited \cr
LazyLoad: \tab yes\cr }
}
\author{
Devin S. Johnson

Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}

