% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_ikfdnm_likelihood}
\alias{make_ikfdnm_likelihood}
\title{Create IKFDNM model likehood function}
\usage{
make_ikfdnm_likelihood(survival = ~1, recruit = ~1, detection = ~1,
  kf_survival_effects = NULL, fixed_list = NULL, kf_data, dnm_data, N_max,
  ln_prior = NULL)
}
\arguments{
\item{survival}{A model statement for the survival parameters of the N-mixture portion of the model}

\item{recruit}{A model statement for the recruitment portion of the N-mixture model}

\item{detection}{A model statement for the the detection portion of the N-mixture model}

\item{kf_survival_effects}{A model statement describing the difference between the known-fate survival and the survival of the N-mixture described in
\code{dnm_survival}.}

\item{fixed_list}{A list object with named entries: \code{survival}, \code{recruit}, and \code{det}. Each of the entries is a vector of length
\code{nrow(data)}.}

\item{kf_data}{A data.frame object containing the known fate data for analysis.}

\item{dnm_data}{A data.frame comtaining the dynamic N-mixture data for analysis.}

\item{N_max}{An integer giving the maximum abundance of the N-mixture population.}

\item{ln_prior}{A function of the parameter vector, \code{par}, that returns the log-prior density
evaluated at \code{par}.}
}
\value{
A function of the vector \code{par}
}
\description{
This function assimilates the data and desirted model to create a likelihood function that is a function with only one argument, \code{par},
This likelihood function can teither be optimized via \code{\link{optim}} or placed within an MCMC routine.
}
\details{
The data provided in the \code{data} agument contains rows for each survey within each group. The rows must be forst ordered by group,
then time within group. The data must contain the following \code{colnames}:
\itemize{
\item \code{group} A factor variable indicating the survey group.
\item \code{n} The observed abundance count at each survey time.
\item \code{R} The number of new KF individuals sampled from the population at each time.
}

The \code{kf_survival_effects} model state provides a method for specifying the \emph{difference} in survival between the N-mixtue individuals and the
known-fate individuals. Thus, the known-fate individual survival model is equal to \code{~survival_dnm + kf_survival_effects}.
}
\author{
Devin S. Johnson
}

